+++ vcf/src/vcf/FoundationKit/TextOutputStream.h
@@ -32,13 +32,13 @@ public:
 
 	virtual ~TextOutputStream();
 
-	virtual void seek(const uint32& offset, const SeekType& offsetFrom);
+	virtual void seek(const uint64& offset, const SeekType& offsetFrom);
 
-	virtual uint32 getSize();
+	virtual uint64 getSize();
 
 	virtual char* getBuffer();
 
-	virtual uint32 getCurrentSeekPos() ;
+	virtual uint64 getCurrentSeekPos() ;
 
 	virtual uint32 write( const unsigned char* bytesToRead, uint32 sizeOfBytes );
 
@@ -70,7 +70,7 @@ public:
 	virtual String toString();
 private:
 	String textBuffer_;
-	uint32 size_;
+	uint64 size_;
 	OutputStream* outStream_;
 };
